Refactor: Restructure project for Gradle build
- Migrated build system to Gradle.
- Removed Eclipse project files (.project, .classpath, .settings/) as they are no longer needed.

We now follow conventions better
