+++ .gitignore
@@ -1,1 +1,5 @@
 /bin/
+.env
+.idea
+*.iml
+/.settings/
