+++ src/main/java/com/studyflow/models/LoginResponse.java
@@ -1,8 +1,6 @@
-package models;
+package com.studyflow.models;
 
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
-@JsonIgnoreProperties(ignoreUnknown = true)
 public class LoginResponse {
     public String access_token;
     public String refresh_token;
@@ -10,7 +8,6 @@ public class LoginResponse {
     public int expires_in;
     public User user;
 
-    @JsonIgnoreProperties(ignoreUnknown = true)
     public static class User {
         public String id;
         public String email;
