+++ src/module-info.java
@@ -7,7 +7,7 @@
 module StudyFlow {
     requires java.net.http;
 
-    exports auth;
-    exports models;
-    exports utils;
+    exports main.java.com.studyflow.service.auth;
+    exports main.java.com.studyflow.models;
+    exports main.java.com.studyflow.utils;
 }
